Create CommentMutationStore type
