+++ src/components/CommentCard/Actions.tsx
@@ -8,6 +8,7 @@ import { Popover, PopoverContent, PopoverTrigger } from '@nextui-org/react';
 
 import { useDeleteComment } from '@/lib/hooks';
 import { useCommentMutationStore } from '@/lib/stores';
+import { type CommentMutationStore } from '@/lib/types';
 
 import { Button } from '../Button';
 
@@ -19,7 +20,11 @@ const Actions = (): ReactNode => {
   const { mutate: deleteComment } = useDeleteComment();
 
   const setCommentMutationId: (id: number | undefined) => void =
-    useCommentMutationStore((state) => state.setId);
+    useCommentMutationStore(
+      (state: CommentMutationStore): ((id: number | undefined) => void) => {
+        return state.setId;
+      }
+    );
 
   const handleClick = (): void => {
     deleteComment(comment?.id, {
