+++ src/lib/types/CommentMutationStore.ts
@@ -1,0 +1,10 @@
+import type CommentSchema from './CommentSchema';
+
+type CommentMutationStore = {
+  data: CommentSchema | null;
+  id: number | undefined;
+  setData: (data: CommentSchema) => void;
+  setId: (id: number | undefined) => void;
+};
+
+export default CommentMutationStore;
