+++ src/lib/types/index.ts
@@ -6,6 +6,7 @@ export { type default as PostSchema } from './PostSchema';
 export { type default as ReactionSchema } from './ReactionSchema';
 export { type default as ViewSchema } from './ViewSchema';
 
+export { type default as CommentMutationStore } from './CommentMutationStore';
 export { type default as PostMutationStore } from './PostMutationStore';
 
 export { type default as DummyJSONComment } from './DummyJSONComment';
